refactor(P4b-3): extract ContextWindow mixin verbatim from Executor
Functional message-list transformations — empty-assistant pruning,
thinking-block strip, oversized-tool-result trimming, and context-
window-aware compaction — move into Inference::Executor::ContextWindow.
Verbatim move; THINKING_TAG_PAIRS resolves through the surrounding
class via lexical scope.

Executor.rb: 1827 → 1681 lines. context_window.rb: 163 lines.
